Extract invoicing from OfferCard to InvoiceCard, rename to BerthOfferCard
