+++ src/features/applicationView/berthOfferCard/__fixtures__/mockData.ts
