+++ src/features/applicationView/berthOfferCard/__tests__/utils.test.ts
