+++ src/features/applicationView/berthOfferCard/fragments.ts
