+++ src/features/applicationView/queries.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import { BERTH_LEASE_FRAGMENT, ADDITIONAL_SERVICES_FRAGMENT } from './offerCard/fragments';
+import { BERTH_LEASE_FRAGMENT, ADDITIONAL_SERVICES_FRAGMENT } from './berthOfferCard/fragments';
 
 export const INDIVIDUAL_APPLICATION_QUERY = gql`
   query INDIVIDUAL_APPLICATION($id: ID!) {
