+++ src/features/invoiceCard/types.ts
@@ -1,0 +1,26 @@
+import React from 'react';
+
+import { ProductServiceType } from '../../@types/__generated__/globalTypes';
+import { IconProps } from '../../common/icons';
+
+export interface Product {
+  id: string;
+  name: ProductServiceType;
+  price: number;
+  orderId: string;
+}
+
+export interface Order {
+  id: string;
+  orderNumber: string;
+  price: number;
+  totalPrice: number;
+  fixedProducts: Product[];
+  optionalProducts: Product[];
+}
+
+export type PlaceProperty = {
+  prop: boolean | null;
+  key: string;
+  icon: (props: IconProps) => React.ReactElement | null;
+};
