+++ src/locales/fi.json
@@ -249,15 +249,6 @@
       "phone": "Puhelinnumero",
       "email": "Sähköpostiosoite",
       "comment": "Huomiot"
-    },
-    "sendInvoice": {
-      "title": "Lähetä lasku",
-      "instructions": {
-        "paragraph1": "Lasku lähetetään asiakkaan sähköpostiosoitteeseen ({{email}}) ja siihen merkitään alla oleva eräpäivä.",
-        "paragraph2": "Lähetettyä laskua ei voi enää muuttaa, mutta sen voi perua. Perumisesta lähetetään ilmoitus asiakkaalle"
-      },
-      "missingEmail": "Sähköpostiosoite puuttuu",
-      "dueDate": "Laskun eräpäivä"
     }
   },
   "harborList": {
@@ -684,6 +675,15 @@
     "placeDetails": {
       "title": "Paikan tiedot",
       "maintenanceDetails": "Huoltotiedot"
+    },
+    "sendInvoice": {
+      "title": "Lähetä lasku",
+      "instructions": {
+        "paragraph1": "Lasku lähetetään asiakkaan sähköpostiosoitteeseen ({{email}}) ja siihen merkitään alla oleva eräpäivä.",
+        "paragraph2": "Lähetettyä laskua ei voi enää muuttaa, mutta sen voi perua. Perumisesta lähetetään ilmoitus asiakkaalle"
+      },
+      "missingEmail": "Sähköpostiosoite puuttuu",
+      "dueDate": "Laskun eräpäivä"
     }
   }
 }
